app/routers/document.py
from fastapi import APIRouter, File, UploadFile, Depends
from sqlalchemy.orm import Session
from app.database import SessionLocal
from app.models import Document
from app.schemas import DocumentResponse
import shutil, os
from dotenv import load_dotenv

load_dotenv()

UPLOAD_DIR = os.getenv("UPLOAD_DIR", "./uploads")
os.makedirs(UPLOAD_DIR, exist_ok=True)

router = APIRouter()

def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()

@router.post("/upload-document", response_model=DocumentResponse)
async def upload_file(file: UploadFile = File(...), db: Session = Depends(get_db)):
    # Save file to disk
    file_location = os.path.join(UPLOAD_DIR, file.filename)
    with open(file_location, "wb") as f:
        shutil.copyfileobj(file.file, f)

    document = Document(filename=file.filename, path=file_location)
    db.add(document)
    db.commit()
    db.refresh(document)

    return document

